More parse fixes - Fix template expression and type annotation issue - Combine import and export part - Move number and options out of parser/lib.rs
