+++ checker/specification/to_implement.md
@@ -830,19 +830,6 @@ func(cb => { cb satisfies boolean }, "hi")
 
 > Was working, now broken (or removed)
 
-#### Readonly property
-
-> Should be working but parser current wraps `a` as `Readonly<string>` :(
-
-```ts
-function x(p: { readonly a: string, b: string }) {
-    p.a = "hi";
-	p.b = "hi";
-}
-```
-
-- Cannot write to property 'a'
-
 #### Destructuring using iterator
 
 ```ts
@@ -887,21 +874,6 @@ function x(a: GreaterThan<4>) {
 ```
 
 - Expected string, found boolean
-
-#### Tagged template literal
-
-> Waiting for parser definition updated to make this easier
-
-```ts
-function myTag(static_parts: Array<string>, other: string) {
-	return { static_parts, other }
-}
-
-const name = "Ben";
-myTag`${name}Hello ` satisfies string
-```
-
-- Expected string, found { static_parts: ["", "Hello "], other: "Ben" }
 
 ### Control flow
