+++ checker/src/synthesis/hoisting.rs
@@ -188,7 +188,7 @@ pub(crate) fn hoist_statements<T: crate::ReadFromFS>(
 						parser::declarations::import::ImportedItems::Parts(parts) => {
 							crate::utilities::notify!("{:?}", parts);
 							crate::features::modules::ImportKind::Parts(
-								parts.iter().flatten().filter_map(import_part_to_name_pair),
+								parts.iter().flatten().filter_map(part_to_name_pair),
 							)
 						}
 						parser::declarations::import::ImportedItems::All { under } => match under {
@@ -263,7 +263,7 @@ pub(crate) fn hoist_statements<T: crate::ReadFromFS>(
 						},
 					..
 				}) => {
-					let parts = parts.iter().filter_map(export_part_to_name_pair);
+					let parts = parts.iter().filter_map(part_to_name_pair);
 
 					import_items(
 						environment,
@@ -842,68 +842,29 @@ pub(crate) fn hoist_statements<T: crate::ReadFromFS>(
 	}
 }
 
-fn import_part_to_name_pair(item: &parser::declarations::ImportPart) -> Option<NamePair<'_>> {
-	match item {
-		parser::declarations::ImportPart::Name(name) => {
-			if let VariableIdentifier::Standard(name, position) = name {
-				Some(NamePair { value: name, r#as: name, position: *position })
-			} else {
-				None
-			}
-		}
-		parser::declarations::ImportPart::NameWithAlias { name, alias, position } => {
-			Some(NamePair {
-				value: match alias {
-					parser::declarations::ImportExportName::Reference(item)
-					| parser::declarations::ImportExportName::Quoted(item, _) => item,
-					parser::declarations::ImportExportName::Marker(_) => {
-						// TODO I think okay
-						return None;
-					}
-				},
-				r#as: name,
-				position: *position,
-			})
-		}
-		parser::declarations::ImportPart::PrefixComment(_, item, _) => {
-			item.as_deref().and_then(import_part_to_name_pair)
-		}
-		parser::declarations::ImportPart::PostfixComment(item, _, _) => {
-			import_part_to_name_pair(item)
-		}
-	}
-}
-
-pub(super) fn export_part_to_name_pair(
-	item: &parser::declarations::export::ExportPart,
+// TODO with `type`
+pub(super) fn part_to_name_pair<T: parser::declarations::ImportOrExport>(
+	item: &parser::declarations::ImportExportPart<T>,
 ) -> Option<NamePair<'_>> {
-	match item {
-		parser::declarations::export::ExportPart::Name(name) => {
-			if let VariableIdentifier::Standard(name, position) = name {
-				Some(NamePair { value: name, r#as: name, position: *position })
-			} else {
-				None
+	if let VariableIdentifier::Standard(ref name, position) = item.name {
+		let value = match &item.alias {
+			Some(
+				parser::declarations::ImportExportName::Reference(item)
+				| parser::declarations::ImportExportName::Quoted(item, _),
+			) => item,
+			Some(parser::declarations::ImportExportName::Marker(_)) => {
+				// TODO I think okay
+				return None;
 			}
+			None => name,
+		};
+		if T::PREFIX {
+			Some(NamePair { value, r#as: name, position })
+		} else {
+			Some(NamePair { value: name, r#as: value, position })
 		}
-		parser::declarations::export::ExportPart::NameWithAlias { name, alias, position } => {
-			Some(NamePair {
-				value: name,
-				r#as: match alias {
-					parser::declarations::ImportExportName::Reference(item)
-					| parser::declarations::ImportExportName::Quoted(item, _) => item,
-					parser::declarations::ImportExportName::Marker(_) => {
-						return None;
-					}
-				},
-				position: *position,
-			})
-		}
-		parser::declarations::export::ExportPart::PrefixComment(_, item, _) => {
-			item.as_deref().and_then(export_part_to_name_pair)
-		}
-		parser::declarations::export::ExportPart::PostfixComment(item, _, _) => {
-			export_part_to_name_pair(item)
-		}
+	} else {
+		None
 	}
 }
