+++ parser/src/declarations/import.rs
@@ -1,23 +1,20 @@
-use get_field_by_type::GetFieldByType;
 use iterator_endiate::EndiateIteratorExt;
 use source_map::Span;
 use tokenizer_lib::{sized_tokens::TokenStart, Token, TokenReader};
 
 use crate::{
-	ast::object_literal::ObjectLiteral, derive_ASTNode, errors::parse_lexing_error,
-	parse_bracketed, throw_unexpected_token, tokens::token_as_identifier, ASTNode, ListItem,
-	Marker, ParseOptions, ParseResult, ParsingState, Quoted, TSXKeyword, TSXToken,
-	VariableIdentifier,
+	ast::object_literal::ObjectLiteral, derive_ASTNode, parse_bracketed, throw_unexpected_token,
+	ASTNode, ParseOptions, ParseResult, ParsingState, TSXKeyword, TSXToken, VariableIdentifier,
 };
 use visitable_derive::Visitable;
 
-use super::ImportLocation;
+use super::{ImportExportPart, ImportLocation};
 
 /// Side effects is represented under the Parts variant where the vector is empty
 #[derive(Debug, Clone, PartialEq, Visitable)]
 #[apply(derive_ASTNode)]
 pub enum ImportedItems {
-	Parts(Option<Vec<ImportPart>>),
+	Parts(Option<Vec<ImportExportPart<ImportDeclaration>>>),
 	All { under: VariableIdentifier },
 }
 
@@ -36,48 +33,6 @@ pub struct ImportDeclaration {
 	pub position: Span,
 	#[cfg(feature = "extras")]
 	pub reversed: bool,
-}
-
-/// TODO `default` should have its own variant?
-#[derive(Debug, Clone, PartialEq)]
-#[apply(derive_ASTNode)]
-pub enum ImportExportName {
-	Reference(String),
-	Quoted(String, Quoted),
-	/// For typing here
-	#[cfg_attr(feature = "self-rust-tokenize", self_tokenize_field(0))]
-	Marker(
-		#[cfg_attr(target_family = "wasm", tsify(type = "Marker<ImportExportName>"))] Marker<Self>,
-	),
-}
-
-impl ImportExportName {
-	pub(crate) fn from_reader(
-		reader: &mut impl TokenReader<TSXToken, crate::TokenStart>,
-		state: &mut crate::ParsingState,
-		options: &ParseOptions,
-	) -> ParseResult<(Self, source_map::End)> {
-		if let Some(Token(TSXToken::Comma, pos)) = reader.peek() {
-			let marker = state.new_partial_point_marker(*pos);
-			return Ok((ImportExportName::Marker(marker), pos.get_end_after(0)));
-		}
-		let token = reader.next().unwrap();
-		if let Token(TSXToken::StringLiteral(alias, quoted), start) = token {
-			let with_length = start.get_end_after(alias.len() + 1);
-			state.constant_imports.push(alias.clone());
-			Ok((ImportExportName::Quoted(alias, quoted), with_length))
-		} else {
-			let (ident, pos) = token_as_identifier(token, "import alias")?;
-			if options.interpolation_points && ident == crate::marker::MARKER {
-				Ok((
-					ImportExportName::Marker(state.new_partial_point_marker(pos.get_start())),
-					pos.get_end(),
-				))
-			} else {
-				Ok((ImportExportName::Reference(ident), pos.get_end()))
-			}
-		}
-	}
 }
 
 impl ASTNode for ImportDeclaration {
@@ -274,7 +229,7 @@ pub(crate) fn parse_import_specifier_and_parts(
 		let end = under.get_position().get_end();
 		(ImportedItems::All { under }, end)
 	} else if let Some(Token(TSXToken::OpenBrace, _)) = peek {
-		let (parts, _, end) = parse_bracketed::<ImportPart>(
+		let (parts, _, end) = parse_bracketed::<ImportExportPart<_>>(
 			reader,
 			state,
 			options,
@@ -297,152 +252,4 @@ pub(crate) fn parse_import_specifier_and_parts(
 		items,
 		end,
 	})
-}
-
-/// <https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Statements/import#syntax>
-#[apply(derive_ASTNode)]
-#[derive(Debug, Clone, PartialEq, Visitable, GetFieldByType)]
-#[get_field_by_type_target(Span)]
-pub enum ImportPart {
-	Name(VariableIdentifier),
-	NameWithAlias {
-		name: String,
-		alias: ImportExportName,
-		position: Span,
-	},
-	PrefixComment(
-		String,
-		#[cfg_attr(target_family = "wasm", tsify(type = "ImportPart | null"))] Option<Box<Self>>,
-		Span,
-	),
-	PostfixComment(
-		#[cfg_attr(target_family = "wasm", tsify(type = "ImportPart"))] Box<Self>,
-		String,
-		Span,
-	),
-}
-
-impl ListItem for ImportPart {
-	type LAST = ();
-}
-
-impl ASTNode for ImportPart {
-	fn get_position(&self) -> Span {
-		*GetFieldByType::get(self)
-	}
-
-	// TODO also single line comments here
-	fn from_reader(
-		reader: &mut impl TokenReader<TSXToken, crate::TokenStart>,
-		state: &mut crate::ParsingState,
-		options: &ParseOptions,
-	) -> ParseResult<Self> {
-		let token = reader.next().ok_or_else(parse_lexing_error)?;
-		if let Token(TSXToken::MultiLineComment(comment), start) = token {
-			let (position, under) =
-				if let Some(Token(TSXToken::CloseBrace | TSXToken::Comma, _)) = reader.peek() {
-					(start.with_length(comment.len() + 2), None)
-				} else {
-					let part = Self::from_reader(reader, state, options)?;
-					(start.union(part.get_position()), Some(Box::new(part)))
-				};
-			Ok(Self::PrefixComment(comment, under, position))
-		} else {
-			let (alias, alias_pos) =
-				if let Token(TSXToken::StringLiteral(alias, quoted), start) = token {
-					let with_length = start.with_length(alias.len() + 2);
-					(ImportExportName::Quoted(alias, quoted), with_length)
-				} else {
-					let (ident, pos) = token_as_identifier(token, "import alias")?;
-					(ImportExportName::Reference(ident), pos)
-				};
-
-			let mut value = match alias {
-				ImportExportName::Quoted(..) => {
-					let _ = state.expect_keyword(reader, TSXKeyword::As)?;
-					let (name, pos) = token_as_identifier(
-						reader.next().ok_or_else(parse_lexing_error)?,
-						"import name",
-					)?;
-					let position = alias_pos.union(pos);
-					Self::NameWithAlias { name, alias, position }
-				}
-				ImportExportName::Reference(reference) => {
-					if state.optionally_expect_keyword(reader, TSXKeyword::As).is_some() {
-						let (name, pos) = token_as_identifier(
-							reader.next().ok_or_else(parse_lexing_error)?,
-							"import name",
-						)?;
-						let position = alias_pos.union(pos);
-						Self::NameWithAlias {
-							name,
-							alias: ImportExportName::Reference(reference),
-							position,
-						}
-					} else {
-						Self::Name(VariableIdentifier::Standard(reference, alias_pos))
-					}
-				}
-				ImportExportName::Marker(id) => Self::Name(VariableIdentifier::Marker(
-					Marker(id.0, Default::default()),
-					alias_pos,
-				)),
-			};
-			while let Some(Token(TSXToken::MultiLineComment(_), _)) = reader.peek() {
-				let Some(Token(TSXToken::MultiLineComment(c), start)) = reader.next() else {
-					unreachable!()
-				};
-				let pos = value.get_position().union(start.get_end_after(c.len() + 2));
-				value = Self::PostfixComment(Box::new(value), c, pos);
-			}
-			Ok(value)
-		}
-	}
-
-	fn to_string_from_buffer<T: source_map::ToString>(
-		&self,
-		buf: &mut T,
-		options: &crate::ToStringOptions,
-		local: crate::LocalToStringInformation,
-	) {
-		match self {
-			ImportPart::Name(name) => {
-				name.to_string_from_buffer(buf, options, local);
-			}
-			ImportPart::NameWithAlias { name, alias, .. } => {
-				match alias {
-					ImportExportName::Reference(alias) => buf.push_str(alias),
-					ImportExportName::Quoted(alias, q) => {
-						buf.push(q.as_char());
-						buf.push_str(alias);
-						buf.push(q.as_char());
-					}
-					ImportExportName::Marker(_) => {}
-				}
-				buf.push_str(" as ");
-				buf.push_str(name);
-			}
-			ImportPart::PrefixComment(comment, inner, _) => {
-				if options.should_add_comment(comment.starts_with('.')) {
-					buf.push_str("/*");
-					buf.push_str(comment);
-					buf.push_str("*/");
-					if inner.is_some() {
-						buf.push(' ');
-					}
-				}
-				if let Some(inner) = inner {
-					inner.to_string_from_buffer(buf, options, local);
-				}
-			}
-			ImportPart::PostfixComment(inner, comment, _) => {
-				inner.to_string_from_buffer(buf, options, local);
-				if options.should_add_comment(comment.starts_with('.')) {
-					buf.push_str("/*");
-					buf.push_str(comment);
-					buf.push_str("*/ ");
-				}
-			}
-		}
-	}
 }
