+++ parser/src/lib.rs
@@ -13,6 +13,8 @@ pub mod generator_helpers;
 mod lexer;
 pub mod marker;
 mod modules;
+pub mod number;
+pub mod options;
 pub mod property_key;
 pub mod statements;
 mod tokens;
@@ -38,6 +40,7 @@ pub use generator_helpers::IntoAST;
 use iterator_endiate::EndiateIteratorExt;
 pub use lexer::{lex_script, LexerOptions};
 pub use modules::Module;
+pub use options::*;
 pub use property_key::PropertyKey;
 pub use source_map::{self, SourceId, Span};
 pub use statements::Statement;
@@ -57,8 +60,6 @@ use tokenizer_lib::{
 };
 
 pub(crate) use tokenizer_lib::sized_tokens::TokenStart;
-
-use std::{borrow::Cow, str::FromStr};
 
 use crate::errors::parse_lexing_error;
 
@@ -89,203 +90,6 @@ impl Quoted {
 			Quoted::Double => '"',
 		}
 	}
-}
-
-/// Options to customize parsing
-#[allow(unused)]
-#[derive(Copy, Clone)]
-// TODO: Can be refactored with bit to reduce memory
-#[allow(clippy::struct_excessive_bools)]
-#[cfg_attr(feature = "serde-serialize", derive(serde::Deserialize), serde(default))]
-#[cfg_attr(target_family = "wasm", derive(tsify::Tsify))]
-pub struct ParseOptions {
-	/// Parsing of [JSX](https://facebook.github.io/jsx/) (includes some additions)
-	pub jsx: bool,
-	/// allow type annotations
-	pub type_annotations: bool,
-	/// just definition file
-	pub type_definition_module: bool,
-	/// Allow custom characters in JSX attributes
-	pub special_jsx_attributes: bool,
-	/// Parses decorators on items
-	pub decorators: bool,
-	/// Skip **all** comments from the AST
-	pub comments: Comments,
-	/// See [`crate::extensions::is_expression::IsExpression`]
-	pub is_expressions: bool,
-	/// Allows functions to be prefixed with 'server'
-	pub custom_function_headers: bool,
-	/// TODO temp for seeing how channel performs
-	pub buffer_size: usize,
-	/// Has no effect on WASM. Increase for deeply nested AST structures
-	pub stack_size: Option<usize>,
-	/// Useful for LSP information
-	pub record_keyword_positions: bool,
-	/// For the generator
-	pub interpolation_points: bool,
-	/// Extra
-	pub destructuring_type_annotation: bool,
-	/// Extra
-	pub extra_operators: bool,
-	/// For formatting
-	pub retain_blank_lines: bool,
-	/// For LSP
-	pub partial_syntax: bool,
-}
-
-impl ParseOptions {
-	fn get_lex_options(&self) -> LexerOptions {
-		LexerOptions {
-			comments: self.comments,
-			lex_jsx: self.jsx,
-			allow_unsupported_characters_in_jsx_attribute_keys: self.special_jsx_attributes,
-		}
-	}
-
-	#[must_use]
-	pub fn all_features() -> Self {
-		Self {
-			jsx: true,
-			type_annotations: true,
-			type_definition_module: false,
-			special_jsx_attributes: true,
-			comments: Comments::All,
-			decorators: true,
-			custom_function_headers: true,
-			is_expressions: true,
-			buffer_size: 100,
-			stack_size: None,
-			record_keyword_positions: true,
-			// Only used in the AST-generator
-			interpolation_points: false,
-			partial_syntax: true,
-			destructuring_type_annotation: true,
-			extra_operators: true,
-			retain_blank_lines: true,
-		}
-	}
-}
-
-// TODO unsure about some of these defaults, may change in future
-impl Default for ParseOptions {
-	fn default() -> Self {
-		Self {
-			jsx: true,
-			type_annotations: true,
-			type_definition_module: false,
-			special_jsx_attributes: false,
-			comments: Comments::All,
-			decorators: true,
-			custom_function_headers: false,
-			is_expressions: false,
-			buffer_size: 100,
-			stack_size: None,
-			record_keyword_positions: false,
-			interpolation_points: false,
-			partial_syntax: false,
-			destructuring_type_annotation: false,
-			extra_operators: false,
-			retain_blank_lines: false,
-		}
-	}
-}
-
-/// Settings for serializing `ASTNodes`
-// TODO: Can be refactored with bit to reduce memory
-#[allow(clippy::struct_excessive_bools)]
-#[cfg_attr(feature = "serde-serialize", derive(serde::Deserialize), serde(default))]
-#[cfg_attr(target_family = "wasm", derive(tsify::Tsify))]
-pub struct ToStringOptions {
-	/// Does not include whitespace minification
-	pub pretty: bool,
-	/// Blocks have trailing semicolons. Has no effect if pretty == false
-	pub trailing_semicolon: bool,
-	/// Single statements get put on the same line as their parent statement
-	pub single_statement_on_new_line: bool,
-	/// Include type annotations (and additional TypeScript) syntax
-	pub include_type_annotations: bool,
-	/// TODO unsure about this
-	pub include_decorators: bool,
-	pub comments: Comments,
-	pub indent_with: String,
-	/// If false, panics if sees JSX
-	pub expect_jsx: bool,
-	/// For partial AST, marker nodes may exist. This allows pretty printing on invalid source
-	/// but should be `false` for builds
-	///
-	/// if `false` and a marker node is found, printing will panic
-	pub expect_markers: bool,
-	/// has no effect under !pretty
-	pub max_line_length: u8,
-}
-
-impl Default for ToStringOptions {
-	fn default() -> Self {
-		ToStringOptions {
-			pretty: true,
-			include_type_annotations: false,
-			single_statement_on_new_line: true,
-			include_decorators: false,
-			comments: Comments::All,
-			expect_jsx: false,
-			trailing_semicolon: false,
-			expect_markers: false,
-			indent_with: "\t".to_owned(),
-			max_line_length: u8::MAX,
-		}
-	}
-}
-
-impl ToStringOptions {
-	#[must_use]
-	pub fn minified() -> Self {
-		ToStringOptions {
-			pretty: false,
-			comments: Comments::None,
-			indent_with: String::new(),
-			..Default::default()
-		}
-	}
-
-	/// With TypeScript type syntax
-	#[must_use]
-	pub fn typescript() -> Self {
-		ToStringOptions { include_type_annotations: true, ..Default::default() }
-	}
-
-	/// Whether to include comment in source
-	pub(crate) fn should_add_comment(&self, is_document_comment: bool) -> bool {
-		matches!(self.comments, Comments::All)
-			|| (matches!(self.comments, Comments::JustDocumentation) && is_document_comment)
-	}
-
-	pub(crate) fn add_indent<T: source_map::ToString>(&self, indent: u8, buf: &mut T) {
-		if self.pretty {
-			(0..indent).for_each(|_| buf.push_str(&self.indent_with));
-		}
-	}
-
-	/// Adds whitespace **conditionally** (based on pretty setting)
-	pub(crate) fn push_gap_optionally<T: source_map::ToString>(&self, buf: &mut T) {
-		if self.pretty {
-			buf.push(' ');
-		}
-	}
-
-	pub(crate) fn enforce_limit_length_limit(&self) -> bool {
-		self.pretty && self.max_line_length != u8::MAX
-	}
-}
-
-#[derive(Debug, Default, Clone, Copy)]
-#[cfg_attr(feature = "serde-serialize", derive(serde::Deserialize))]
-#[cfg_attr(target_family = "wasm", derive(tsify::Tsify))]
-pub enum Comments {
-	#[default]
-	All,
-	/// Only multiline comments starting with `/**`
-	JustDocumentation,
-	None,
 }
 
 #[derive(Debug, Clone, Copy)]
@@ -544,294 +348,6 @@ impl KeywordPositions {
 
 	fn new() -> Self {
 		Self(Default::default())
-	}
-}
-
-#[derive(Debug, PartialEq, Eq, Clone, Copy, Hash)]
-#[apply(derive_ASTNode)]
-pub enum NumberSign {
-	/// Also implies non negative/missing
-	Positive,
-	Negative,
-}
-
-impl NumberSign {
-	pub fn apply<T: std::ops::Neg<Output = T>>(&self, x: T) -> T {
-		match self {
-			NumberSign::Positive => x,
-			NumberSign::Negative => -x,
-		}
-	}
-}
-
-impl std::ops::Neg for NumberSign {
-	type Output = Self;
-
-	fn neg(self) -> Self::Output {
-		match self {
-			NumberSign::Positive => NumberSign::Negative,
-			NumberSign::Negative => NumberSign::Positive,
-		}
-	}
-}
-
-impl std::fmt::Display for NumberSign {
-	fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-		if matches!(self, Self::Negative) {
-			f.write_str("-")
-		} else {
-			Ok(())
-		}
-	}
-}
-
-/// Some of these can't be parsed, but are there to make so that a number expression can be generated from a f64
-///
-/// <https://tc39.es/ecma262/multipage/ecmascript-language-lexical-grammar.html#sec-literals-numeric-literals>
-#[derive(Debug, Clone)]
-#[apply(derive_ASTNode)]
-pub enum NumberRepresentation {
-	Infinity,
-	NegativeInfinity,
-	NaN,
-	Hex { sign: NumberSign, value: u64 },
-	Bin { sign: NumberSign, value: u64 },
-	Octal { sign: NumberSign, value: u64 },
-	Number(f64),
-	Exponential { sign: NumberSign, value: f64, exponent: i32 },
-	BigInt(NumberSign, String),
-}
-
-impl std::hash::Hash for NumberRepresentation {
-	fn hash<H: std::hash::Hasher>(&self, state: &mut H) {
-		core::mem::discriminant(self).hash(state);
-	}
-}
-
-impl TryFrom<NumberRepresentation> for f64 {
-	// BigInt!!
-	type Error = ();
-
-	fn try_from(this: NumberRepresentation) -> Result<Self, Self::Error> {
-		match this {
-			NumberRepresentation::Infinity => Ok(f64::INFINITY),
-			NumberRepresentation::NegativeInfinity => Ok(f64::NEG_INFINITY),
-			NumberRepresentation::NaN => Ok(f64::NAN),
-			NumberRepresentation::Number(value) => Ok(value),
-			NumberRepresentation::Hex { sign, value, .. }
-			| NumberRepresentation::Bin { sign, value, .. }
-			| NumberRepresentation::Octal { sign, value, .. } => {
-				// TODO `value as f64` can lose information? If so should return f64::INFINITY
-				#[allow(clippy::cast_precision_loss)]
-				Ok(sign.apply(value as f64))
-			}
-			NumberRepresentation::Exponential { sign, value, exponent } => {
-				Ok(sign.apply(value * 10f64.powi(exponent)))
-			}
-			NumberRepresentation::BigInt(..) => Err(()),
-		}
-	}
-}
-
-// For code generation
-impl From<f64> for NumberRepresentation {
-	fn from(value: f64) -> Self {
-		if value == f64::INFINITY {
-			Self::Infinity
-		} else if value == f64::NEG_INFINITY {
-			Self::NegativeInfinity
-		} else if value.is_nan() {
-			Self::NaN
-		} else {
-			Self::Number(value)
-		}
-	}
-}
-
-impl FromStr for NumberRepresentation {
-	type Err = String;
-
-	fn from_str(s: &str) -> Result<Self, Self::Err> {
-		if s == "NaN" {
-			return Ok(Self::NaN);
-		}
-
-		if s.contains('_') {
-			return s.replace('_', "").parse();
-		}
-
-		let (sign, s) = if let Some(s) = s.strip_prefix('-') {
-			(NumberSign::Negative, s)
-		} else {
-			(NumberSign::Positive, s)
-		};
-
-		let s = if s.contains('_') { Cow::Owned(s.replace('_', "")) } else { Cow::Borrowed(s) };
-
-		if let Some(s) = s.strip_suffix('n') {
-			Ok(NumberRepresentation::BigInt(sign, s.to_owned()))
-		} else if let Some(s) = s.strip_prefix('0') {
-			let next_char = s.chars().next();
-			match next_char {
-				Some('.') => {
-					if s.len() == 1 {
-						Ok(Self::Number(0f64))
-					} else {
-						Ok(Self::Number(sign.apply(s.parse().map_err(|_| s.to_owned())?)))
-					}
-				}
-				Some('X' | 'x') => {
-					let mut value = 0u64;
-					for c in s[1..].as_bytes() {
-						value <<= 4; // 16=2^4
-						match c {
-							b'0'..=b'9' => {
-								value += u64::from(c - b'0');
-							}
-							b'a'..=b'f' => {
-								value += u64::from(c - b'a') + 10;
-							}
-							b'A'..=b'F' => {
-								value += u64::from(c - b'A') + 10;
-							}
-							_ => return Err(s.to_owned()),
-						}
-					}
-					Ok(Self::Hex { sign, value })
-				}
-				Some('B' | 'b') => {
-					let mut value = 0u64;
-					for c in s[1..].as_bytes() {
-						value <<= 1;
-						match c {
-							b'0' | b'1' => {
-								value += u64::from(c - b'0');
-							}
-							_ => return Err(s.to_owned()),
-						}
-					}
-					Ok(Self::Bin { sign, value })
-				}
-				Some('e' | 'E') => {
-					// Lol
-					let exponent: i32 = s[1..].parse().map_err(|_| s.to_owned())?;
-					Ok(Self::Exponential { sign, value: 0f64, exponent })
-				}
-				// 'o' | 'O' but can also be missed
-				Some(c) => {
-					let uses_character = matches!(c, 'o' | 'O');
-
-					if !uses_character && s.contains(['8', '9', '.']) {
-						return Ok(Self::Number(sign.apply(s.parse().map_err(|_| s.to_owned())?)));
-					}
-
-					// If it uses the the character then skip one, else skip zero
-					let start: usize = uses_character.into();
-
-					let mut value = 0u64;
-					for c in s[start..].as_bytes() {
-						value <<= 3; // 8=2^3
-						if matches!(c, b'0'..=b'7') {
-							value += u64::from(c - b'0');
-						} else {
-							return Err(s.to_owned());
-						}
-					}
-					Ok(Self::Octal { sign, value })
-				}
-				None => Ok(Self::Number(0f64)),
-			}
-		} else if s.starts_with('.') {
-			let value: f64 = format!("0{s}").parse().map_err(|_| s.clone())?;
-			Ok(Self::Number(sign.apply(value)))
-		} else if let Some(s) = s.strip_suffix('.') {
-			Ok(Self::Number(sign.apply(s.parse::<f64>().map_err(|_| s)?)))
-		} else if let Some((left, right)) = s.split_once(['e', 'E']) {
-			let value = left.parse::<f64>().map_err(|_| s.clone())?;
-			if let Ok(exponent) = right.parse::<i32>() {
-				Ok(Self::Exponential { sign, value, exponent })
-			} else if right.starts_with('-') || value == 0f64 {
-				// lol
-				Ok(Self::Number(0f64))
-			} else {
-				Ok(Self::Infinity)
-			}
-		} else {
-			Ok(Self::Number(sign.apply(s.parse::<f64>().map_err(|_| s.clone())?)))
-		}
-	}
-}
-
-impl std::fmt::Display for NumberRepresentation {
-	fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-		write!(f, "{}", self.clone().as_js_string())
-	}
-}
-
-// TODO not great
-impl PartialEq for NumberRepresentation {
-	fn eq(&self, other: &Self) -> bool {
-		if let (Ok(a), Ok(b)) = (f64::try_from(self.clone()), f64::try_from(other.clone())) {
-			a == b
-		} else {
-			// TODO ...
-			false
-		}
-	}
-}
-
-impl Eq for NumberRepresentation {}
-
-impl std::ops::Neg for NumberRepresentation {
-	type Output = Self;
-
-	fn neg(self) -> Self::Output {
-		match self {
-			NumberRepresentation::Infinity => NumberRepresentation::NegativeInfinity,
-			NumberRepresentation::NegativeInfinity => NumberRepresentation::Infinity,
-			NumberRepresentation::NaN => NumberRepresentation::NaN,
-			NumberRepresentation::Hex { sign, value } => {
-				NumberRepresentation::Hex { sign: sign.neg(), value }
-			}
-			NumberRepresentation::Bin { sign, value } => {
-				NumberRepresentation::Bin { sign: sign.neg(), value }
-			}
-			NumberRepresentation::Octal { sign, value } => {
-				NumberRepresentation::Octal { sign: sign.neg(), value }
-			}
-			NumberRepresentation::Number(n) => NumberRepresentation::Number(n.neg()),
-			NumberRepresentation::Exponential { sign, value, exponent } => {
-				NumberRepresentation::Exponential { sign: sign.neg(), value, exponent }
-			}
-			NumberRepresentation::BigInt(sign, value) => {
-				NumberRepresentation::BigInt(sign.neg(), value)
-			}
-		}
-	}
-}
-
-impl NumberRepresentation {
-	#[must_use]
-	pub fn as_js_string(self) -> String {
-		match self {
-			NumberRepresentation::Infinity => "Infinity".to_owned(),
-			NumberRepresentation::NegativeInfinity => "-Infinity".to_owned(),
-			NumberRepresentation::NaN => "NaN".to_owned(),
-			NumberRepresentation::Hex { sign, value, .. } => {
-				format!("{sign}0x{value:x}")
-			}
-			NumberRepresentation::Bin { sign, value, .. } => {
-				format!("{sign}0b{value:b}")
-			}
-			NumberRepresentation::Octal { sign, value } => {
-				format!("{sign}0o{value:o}")
-			}
-			NumberRepresentation::Number(value) => value.to_string(),
-			NumberRepresentation::Exponential { sign, value, exponent } => {
-				format!("{sign}{value}e{exponent}")
-			}
-			NumberRepresentation::BigInt(s, value) => format!("{s}{value}n"),
-		}
 	}
 }
 
@@ -1260,9 +776,10 @@ pub mod ast {
 			FunctionBase, FunctionBody, FunctionHeader, FunctionParameters, MethodHeader,
 			Parameter, ParameterData, SpreadParameter,
 		},
+		number::NumberRepresentation,
 		statements::*,
-		Block, Decorated, ExpressionPosition, NumberRepresentation, PropertyKey,
-		StatementOrDeclaration, StatementPosition, VariableField, VariableIdentifier, WithComment,
+		Block, Decorated, ExpressionPosition, PropertyKey, StatementOrDeclaration,
+		StatementPosition, VariableField, VariableIdentifier, WithComment,
 	};
 
 	pub use source_map::{BaseSpan, SourceId};
